[Jason]refactor the foler structure in security logic
