+++ src/main/java/com/tw/apistack/security/api/LoginController.java
@@ -1,5 +1,10 @@
-package com.tw.apistack.security;
+package com.tw.apistack.security.api;
 
+import com.tw.apistack.security.api.dto.NewPassword;
+import com.tw.apistack.security.authentication.JwtToken;
+import com.tw.apistack.security.authentication.TokenAuthenticationService;
+import com.tw.apistack.security.authrization.ApiStackAuthenticationSuccessHandler;
+import com.tw.apistack.security.authrization.ApiStackUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.User;
