+++ src/main/java/com/tw/apistack/security/authentication/JwtToken.java
@@ -1,4 +1,4 @@
-package com.tw.apistack.security;
+package com.tw.apistack.security.authentication;
 
 public class JwtToken {
     private String token;
