+++ src/main/java/com/tw/apistack/security/authrization/ApiStackAuthenticationFailureHandler.java
@@ -1,4 +1,4 @@
-package com.tw.apistack.security;
+package com.tw.apistack.security.authrization;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
