+++ src/main/java/com/tw/apistack/security/authrization/ApiStackUserDetailsService.java
@@ -1,7 +1,11 @@
 /* Starbucks Licensed  */
 
-package com.tw.apistack.security;
+package com.tw.apistack.security.authrization;
 
+import com.tw.apistack.security.exception.BadPasswordException;
+import com.tw.apistack.security.api.dto.NewPassword;
+import com.tw.apistack.security.core.model.User;
+import com.tw.apistack.security.core.respository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
