+++ src/main/java/com/tw/apistack/security/core/model/Authorities.java
@@ -1,4 +1,4 @@
-package com.tw.apistack.security;
+package com.tw.apistack.security.core.model;
 
 public enum Authorities {
     ROLE_USER, ROLE_ADMIN
