+++ src/main/java/com/tw/apistack/security/core/model/User.java
@@ -1,4 +1,4 @@
-package com.tw.apistack.security;
+package com.tw.apistack.security.core.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
