+++ src/main/java/com/tw/apistack/security/core/respository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.tw.apistack.security;
+package com.tw.apistack.security.core.respository;
 
+import com.tw.apistack.security.core.model.User;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
