+++ src/main/java/com/tw/apistack/security/exception/BadPasswordException.java
@@ -1,4 +1,4 @@
-package com.tw.apistack.security;
+package com.tw.apistack.security.exception;
 
 public class BadPasswordException extends RuntimeException {
     public BadPasswordException(String passsword) {
