+++ src/test/java/com/tw/apistack/security/LoginControllerIT.java
@@ -1,5 +1,6 @@
 package com.tw.apistack.security;
 
+import com.tw.apistack.security.core.model.User;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
